requirements.txt
streamlit
pandas
geopy
streamlit-mic-recorder
google-generativeai